% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{adeg}
\alias{adeg}
\title{TQT Study ADEG data.}
\format{
\describe{
\item{STUDYID}{Study Identifier, Char}
\item{USUBJID}{Unique Subject Identifier, Char}
\item{TRTSEQP}{Planned Sequence of Treatments, Char}
\item{TRTP}{Planned Treatment, Char}
\item{TRTSEQA}{Actual Sequence of Treatments, Char}
\item{TRTA}{Actual Treatment, Char}
\item{PARAM}{Parameter, Char}
\item{PARAMCD}{Parameter Code, Char}
\item{APERIOD}{Period , Num}
\item{APERIODC}{Period (C), Char}
\item{AVISIT}{Analysis Visit, Char}
\item{AVISITN}{Analysis Visit (N), Num}
\item{ATPT}{Analysis Time Point, Char}
\item{ATPTN}{Analysis Time Point (N), Num}
\item{ATPTREF}{Analysis Reference Time Point, Char}
\item{ARDTM}{Date/Time of Reference, Num}
\item{AARDTM}{Actual Date/Time of Reference, Num}
\item{ARRLT}{Actual Relative Time from Reference Time TimePoint, Num}
\item{NRRLT}{Nominal Relative Time from Reference Time Point, Num}
\item{RRLTU}{Relative Time from Reference Time Point Units, Char}
\item{ADTM}{Analysis Date and Time, Num}
\item{ADT}{Analysis Date, Num}
\item{ATM}{Analysis Time, Num}
\item{ADY}{Analysis Relative Day, Num}
\item{AVISDY}{Analysis Visit Day, Num}
\item{APERDAY}{Analysis Nominal Period Day, Num}
\item{EGREPNUM}{ECG Replicate Number, Num}
\item{AVAL}{Analysis Value, Num}
\item{AVALC}{Analysis Value, Char}
\item{EGSTRESU}{Units of Analysis Value, Char}
\item{DTYPE}{Derivation Type, Char}
\item{ABLFL}{Baseline Record Flag, Char}
\item{AEGBLFL}{Baseline Flag for Records used to Derive the Baseline, Char}
\item{BASE}{Baseline Value, Num}
\item{CHG}{Change from Baseline, Num}
\item{BASETYPE}{Baseline Type, Char}
\item{ACOMPFL}{Comparator Record Flag, Char}
\item{COMP}{Comparator Value, Num}
\item{COMPBASE}{Baseline Value of the Comparator, Num}
\item{COMPCHG}{Change from Baseline in the Comparator, Num}
\item{CCOMPCHG}{Change from Baseline Corrected for the Comparator, Num}
\item{COMPTYPE}{Comparator Type, Char}
\item{ECGPCFL}{ECG Matching PK Flag, Char}
\item{EGSEQ}{Sequence Number, Num}
\item{EGLEAD}{Lead Location Used for Measurement, Char}
\item{EGREFID}{ECG Reference ID, Char}
\item{CRITy}{Analysis Criterion y, Char}
\item{CRITyFL}{Criterion y Evaluation Result Flag, Char}
\item{IUTANLFL}{Intersection Union Test Analysis Flag, Char}
\item{CATANLFL}{Categorical Analysis Flag, Char}
\item{ASEANLFL}{Assay Sensitivity Analysis Flag, Char}
\item{CQTANLFL}{Concentration QT Analysis Flag, Char}
}
}
\source{
https://physionet.org/content/ecgcipa/1.0.0/
}
\usage{
adeg
}
\description{
CiPA ECG Validation Study
See https://physionet.org/content/ecgcipa/1.0.0/
}
\keyword{datasets}
