% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{TQT Study subject level ADSL data.}
\format{
\describe{
\item{STUDYID}{Study Identifier, Char}
\item{USUBJID}{Unique Subject Identifier, Char}
\item{SUBJID}{Subject Identifier for the Study, Char}
\item{SITEID}{Site Identifier, Char}
\item{AGE}{Age, Num}
\item{AGEU}{Age Units, Char}
\item{SEX}{Sex, Char}
\item{RACE}{Race, Char}
\item{ETHNIC}{Ethnic, Char}
\item{ARM}{Planned Arm, Char}
\item{ACTARM}{Actual Arm, Char}
\item{TRT01P}{Planned Treatment for Period 1, Char}
\item{TRT02P}{Planned Treatment for Period 2, Char}
\item{TRT01PN}{Planned Treatment for Period 1 (N), Num}
\item{TRT02PN}{Planned Treatment for Period 2 (N), Num}
\item{TRT01A}{Actual Treatment for Period 1, Char}
\item{TRT02A}{Actual Treatment for Period 2, Char}
\item{TRT01AN}{Actual Treatment for Period 1 (N), Num}
\item{TRT02AN}{Actual Treatment for Period 2 (N), Num}
\item{TRTSEQP}{Planned Sequence of Treatments, Char}
\item{TRTSEQPN}{Planned Sequence of Treatments (N), Num}
\item{TRTSEQA}{Actual Sequence of Treatments, Char}
\item{TRTSEQAN}{Actual Sequence of Treatments (N), Num}
}
}
\source{
https://physionet.org/content/ecgcipa/1.0.0/
}
\usage{
adsl
}
\description{
CiPA ECG Validation Study
See https://physionet.org/content/ecgcipa/1.0.0/
}
\keyword{datasets}
