% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{dm_ph1}
\alias{dm_ph1}
\title{Phase I SAD-MAD Study demographic data.}
\format{
\describe{
\item{PART}{SAD or MAD.}
\item{ID}{Subject identifier.}
\item{TREAT}{A, B, C, D, or E.}
\item{AGE}{Subjects age in years.}
\item{SEX}{Subjects sex, M=Male, F=Female.}
\item{RACE}{Subjects race.}
}
}
\source{
source data
}
\usage{
dm_ph1
}
\description{

}
\keyword{datasets}
