% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{dm_ph2}
\alias{dm_ph2}
\title{Phase 2b Study demographic data.}
\format{
\describe{
\item{ID}{Subject identifier.}
\item{TREAT}{A, B, C, or D.}
\item{AGE}{Subjects age in years.}
\item{SEX}{Subjects sex, M=Male, F=Female.}
\item{RACE}{Subjects race.}
}
}
\source{
source data
}
\usage{
dm_ph2
}
\description{
Demogrpahic data
}
\keyword{datasets}
