% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{eg_ph1}
\alias{eg_ph1}
\title{Phase I SAD-MAD Study ECG data.}
\format{
\describe{
\item{PART}{SAD or MAD.}
,  \item{ID}{Subject identifier.}
,  \item{DAY}{Day number as in the source data. See DAY2.}
,  \item{DAY2}{For the SAD part, DAY2 = 1 for the day of and the two days following the dose administration, whereas DAY = 1, 2, 3.}
,  \item{VISIT}{Visit number as in the source data. See VISIT2.}
,  \item{VISIT2}{For the SAD part, VISIT2 = 1 for the day of and the two days following the dose administration, whereas VISIT = 1, 2, 3.}
,  \item{TIME}{Time post dose, in hours, reported in the source data. See TIME2.}
,  \item{TIME2}{For the SAD part, TIME is missing for the screening and pre-treatment visits; TIME2 = 0 then. For the SAD part, TIME is missing for the observation on DAY 3; TIME2 = 48 hours then.}
,  \item{PARAM}{HR for heart rate, or QTCF for QTcF.}
,  \item{VALIE}{Observed value of HR, in beats per minute, or QTcF, in milliseconds.}
,  \item{BASE}{Baseline VALUE.}
,  \item{CHANGE}{Change from baseline for observations after the baseline; missing otherwise.}
}
}
\source{
source data
}
\usage{
eg_ph1
}
\description{
Phase I SAD-MAD Study ECG data.
}
\keyword{datasets}
