% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{eg_ph2}
\alias{eg_ph2}
\title{Phase 2b Study ECG data.}
\format{
\describe{
\item{ID}{Subject identifier.}
\item{VISIT}{Visit number.}
\item{DAY}{Day number.}
\item{TIME}{Time relative to dose, or “EARLY WITHDRAWAL”, or missing, as reported in the source data. See TIME2.}
\item{TIME2}{2 if TIME == “2 HOURS POST DOSE”; 0 otherwise. }
\item{PARAM}{HR for heart rate, or QTCF for QTcF.}
\item{VALIE}{Observed value of HR, in beats per minute, or QTcF, in milliseconds.}
\item{BASE}{Baseline VALUE.}
\item{CHANGE}{Change from baseline for observations after the baseline; missing otherwise.}
\item{TREAT}{A, B, C, or D.}
}
}
\source{
source data
}
\usage{
eg_ph2
}
\description{

}
\keyword{datasets}
