% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QT_Central_Tendancy.R
\name{QT_central_tendancy_server}
\alias{QT_central_tendancy_server}
\title{QT Central Tendancy Module - Server}
\usage{
QT_central_tendancy_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with \code{data} and \code{settings} options.}
}
\value{
returns shiny module Server function
}
\description{
QT Central Tendancy Module - Server
}
