% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QT_Central_Tendancy.R
\name{QT_central_tendancy_ui}
\alias{QT_central_tendancy_ui}
\title{QT central tendancy dule - UI}
\usage{
QT_central_tendancy_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
QT central tendancy dule - UI
}
